spectragraph-core/pyproject.toml
[tool.poetry]
name = "spectragraph-core"
version = "0.1.0"
description = "Core utilities and base classes for spectragraph modules"
authors = ["sr-857 <sr-857@users.noreply.github.com>"]
packages = [{ include = "spectragraph_core", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
spectragraph-transforms = { path = "../spectragraph-transforms", develop = true }
neo4j = "^5.0"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
asyncpg = "^0.30"
redis = "^5.0"
celery = "^5.3"
python-dotenv = "^1.0"
requests = "^2.31"
networkx = "^2.6.3"
passlib = {extras = ["bcrypt"], version = "^1.7"}
bcrypt = ">=4.0.0,<5.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3"}
sse-starlette = "^1.8"
alembic = "1.13.0"
phonenumbers = "^9.0.8"
python-multipart = "^0.0.20"
docker = "^7.1.0"
pytest = "^8.4.2"
cryptography = "^45.0.7"
openpyxl = "^3.1"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21"
black = "^23.0"
isort = "^5.12"
flake8 = "^6.0"
mypy = "^1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 